backend/requirements.txt
fastapi[standard]
uvicorn
asyncpg
sqlalchemy
alembic
psycopg2
argon2-cffi